requirements-rag.txt
# Dependencias adicionales para el sistema RAG

# Web Cache y búsqueda
diskcache>=5.6.0  # Cache persistente para búsquedas web
requests>=2.31.0  # Cliente HTTP para SearXNG

# Vector databases (para memoria conversacional)
qdrant-client>=1.7.0  # Qdrant (producción)
chromadb>=0.4.0  # ChromaDB (desarrollo/alternativa)

# Health Dashboard
fastapi>=0.104.0  # Framework web para /health y /metrics
uvicorn>=0.24.0  # ASGI server para FastAPI
jinja2>=3.1.0  # Templates HTML para dashboard

# Opcional: testing adicional
pytest-mock>=3.12.0  # Mocking avanzado para tests
